+++ src/beat_link_trigger/triggers.clj
@@ -41,20 +41,9 @@
            [org.deepsymmetry.electro Metronome]
            [uk.co.xfactorylibrarians.coremidi4j CoreMidiDeviceProvider]))
 
-(defonce ^{:doc "Provides a space for trigger expressions to store
-  values they want to share across triggers. Visible to other
-  namespaces so that, for example, Show expressions can access them."}
-  expression-globals (atom {}))
-
-;; Make the expression globals conveniently available when compiling
-;; shared functions too.
-(in-ns 'beat-link-trigger.expressions)
-
-#_{:clj-kondo/ignore [:unresolved-namespace]}
-(def globals
-  "The Beat Link Trigger expression globals"
-  beat-link-trigger.triggers/expression-globals)
-(in-ns 'beat-link-trigger.triggers)
+;; This used to be where this was defined, but for compilation order reasons, it has been swapped
+;; into the expressions namespace.
+(refer 'beat-link-trigger.expressions :only '[globals] :rename '{globals expression-globals})
 
 (defonce ^{:private true
            :doc "Holds the trigger window, through which we can access and